chore: update test reporter name format in workflow (#6)
This pull request includes a series of changes to the `jobs:` section in
the `.github/workflows/check.yml` file. The changes involve modifying
the naming convention for unit test results to make them more consistent
and platform-independent.

Changes to unit test result naming:

*
[`.github/workflows/check.yml`](diffhunk://#diff-55f05888fd854d9962bdd4d8ccf23a07b610990f58174397b2586b58f8937428L96-R96):
Updated the `name` field for unit test results by removing the
platform-specific (`linux`) part of the name, making it more generic and
applicable to any platform.
[[1]](diffhunk://#diff-55f05888fd854d9962bdd4d8ccf23a07b610990f58174397b2586b58f8937428L96-R96)
[[2]](diffhunk://#diff-55f05888fd854d9962bdd4d8ccf23a07b610990f58174397b2586b58f8937428L120-R120)
[[3]](diffhunk://#diff-55f05888fd854d9962bdd4d8ccf23a07b610990f58174397b2586b58f8937428L147-R147)
[[4]](diffhunk://#diff-55f05888fd854d9962bdd4d8ccf23a07b610990f58174397b2586b58f8937428L174-R174)
[[5]](diffhunk://#diff-55f05888fd854d9962bdd4d8ccf23a07b610990f58174397b2586b58f8937428L198-R198)